requirements.txt
-i https://pypi.org/simple
anyio==4.0.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
exceptiongroup==1.1.3; python_version < '3.11'
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.1; python_version >= '3.8'
httpx==0.25.1; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
install==1.3.5; python_version >= '2.7'
pip==23.3.1; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-telegram-bot==20.6; python_version >= '3.8'
python-telegram-bot-calendar==1.0.5
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.23; python_version >= '3.7'
sqlalchemy-utils==0.41.1; python_version >= '3.6'
typing-extensions==4.8.0; python_version >= '3.8'
